% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotted_line.R
\name{dotted_line}
\alias{dotted_line}
\title{Dotted line
Adds dotted line with label to plot_ly plot}
\usage{
dotted_line(plot, df, x_col, line_val, gluetext, .yshift = 5, .xshift = 2)
}
\arguments{
\item{plot}{plot_ly plot}

\item{df}{Dataframe used to build plot_ly plot}

\item{x_col}{String column name of x-axis}

\item{line_val}{Value on y-axis at which to draw dotted line}

\item{gluetext}{Text fed to glue::glue function used to label dotted line}

\item{.yshift}{How far to shift label relative to y-axis}

\item{.xshift}{How far to shift label relative to x-axis}
}
\value{
plot_ly plot with dotted line and label on dotted line
}
\description{
Dotted line
Adds dotted line with label to plot_ly plot
}
