% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotted_line.R
\name{multi_line}
\alias{multi_line}
\title{Multiple Dotted Lines}
\usage{
multi_line(
  plot,
  df,
  x_col,
  yvec,
  textvec,
  .lineadjust = 0.5,
  .showarrow = T,
  .yshift = 5,
  .xshift = 2
)
}
\arguments{
\item{plot}{Base-plotly chart}

\item{df}{Base-dataframe}

\item{x_col}{String name of x-axis column}

\item{yvec}{Numeric vector of y-values at which to draw dotted lines}

\item{textvec}{Character vector of labels for dotted lines; fed to glue function}

\item{.lineadjust}{How much to adjust the length of the dotted line relative to the x-axis; default 0.5}

\item{.showarrow}{Whether to show an arrow pointing to the dotted line or not; default T}

\item{.yshift}{How much to adjust the y-axis position of the text label relative to the dotted line; default 5}

\item{.xshift}{How much to adjust the x-axis position of the text label relative to the dotted line; default 2}
}
\value{

}
\description{
Create multiple dotted lines in a plot_ly graph
}
\examples{
{

quickplot <- plotly::plot_ly(mtcars,
x = ~ carb,
y = ~ disp,
type = "bar")

multi_line(quickplot, mtcars, "carb", c(200, 250), c("200 line", "250 line"))

}
}
