% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplot_title.R
\name{no_titleaxeslabels}
\alias{no_titleaxeslabels}
\title{Remove Plot and Axes Titles}
\usage{
no_titleaxeslabels(plot)
}
\arguments{
\item{plot}{Plotly plot}
}
\value{
Plotly plot with no labels.
}
\description{
Set title of plotly plot to nothing, and title of axes to nothing. Useful if creating subplots and want to set layout at end.
}
\examples{
{

quickplot <- plotly::plot_ly(mtcars,
x = ~ carb,
y = ~ disp,
type = "bar")

no_titleaxeslabels(quickplot)

}
}
