% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplot_title.R
\name{subplot_title}
\alias{subplot_title}
\title{Function to add a subplot title to a plotly subplot}
\usage{
subplot_title(plot, title, .x = 0.1, .y = 1)
}
\arguments{
\item{plot}{plotly subplot}

\item{title}{title of subplot}

\item{.x}{Numeric position of subplot title relative to x-axis}

\item{.y}{Numeric position of subplot title relative to y-axis}
}
\value{
plotly with subtitle
}
\description{
Function to add a subplot title to a plotly subplot
}
\examples{
{

quickplot <- plotly::plot_ly(mtcars,
x = ~ carb,
y = ~ disp,
type = "bar")

subplot_title(quickplot, "Cars subplot")

}
}
